Fixes unable to send messages
Since socket.io handles events and notifies the other
user when a new message is coming or sent. We don't have
to worry about polling.

The front-end was having an issue of the conversationId not being
defined anytime a message was sent. This was because we were running
the send_message function syncronously rather than asyncrounously and
waiting for the response from the server. That way we have the
data we need (ex. conversationId).