Adds Unread messages to DB and frontend
Frontend:
- Implements read bubble unread indicator
- Sidebar preview text bolding if unreads for user
- A conversation is marked unread when the chat is
opened or a message is recieved through active chat.
- If a converation is opened, it updates the redux store
and marks it all as read while sending a post request to
the backend to update the DB accordingly.
- If a conversation is opened and a message is recieved via
active chat, we only need to update the DB to reset to 0. This is
because anytime a message is recieved we presume
the other hasn't read it and increment the unread count in the DB.

Backend:
- Adds user1 and user2 unread counts to the DB
- Adds setting and /api/conversation POST request handling for
marking a conversation as read.
- Adds proper Sequelize structure to Conversation.