Adds read bubble in messages
Adds read bubble avatar to note where the
other user has last read in the conversation.

the updateMessages reducer was updated to take in a parameter
to choose whether to update the other or our messages. This is needed
when either the other user has read our messages (socket event) and
we want to mark our messages as read. Or when we've read our messages
and want to mark their messages as read.

A socket event update-messages was also added to notify the user in
realtime when we have read their message.