Adds socket security and optimization
Forces the use of our JWT token to authenticate with
our backend socketio middleware on connection creation.
This happens on login and registration.

Also optimizes our onlineUsers to simply use an array and
map the id's to indexes for O(1) get and set operations. Versus
doing more expensive splice and indexOf operations.