Adds message socket security and efficiency
- Messages being sent via sockets were being broadcast
to all sockets. Allowing all users online to see everyones messages
in their socket events. They now send directly only to the socket of
the user intended
- Users could also send messages to others sockets and make it seem
like someone else sent it to them. This was done by changing the
recipientId and conversationId in the "new-message" event.
Checks are done to make sure the user has proper authority to send
to sockets only in their conversations.