converts mapping->hooks
mapStateToProps and mapDispatchToProps are needed
when you have a class component or want to wrap your
component in a Connector.
Using useSelector and useDispatch hooks is a more modern
approach and is easier to implement without a Connector too.

Also prevents errors by adding fixing console issues with
deprecated methods.